feat(F-663): integrate popups in login page (#728)
* feat(F-663): integrate popups in login page

* feat(F-663): remove popup metadata

* feat(F-663): add modal wrapper

* feat(F-663): adjustments

---------
