+++ frontend/messages/de/Login.json
@@ -49,6 +49,26 @@
     "expiredOtpError": "Der Bestätigungscode ist abgelaufen. Bitte fordern Sie einen neuen Code an.",
     "genericError": "Bei der Konto-Bestätigung ist ein Fehler aufgetreten. Bitte versuchen Sie es später noch einmal."
   },
+  "Reset": {
+    "title": "Passwort zurücksetzen",
+    "description": "Bitte geben Sie Ihre E-Mail-Adresse ein, um einen Link zum Zurücksetzen des Passworts zu erhalten.",
+    "sendResetEmailButtonLabel": "Passwort zurücksetzen",
+    "successMessage": "Wenn ein Konto mit dieser E-Mail-Adresse existiert, wurde ein Link zum Zurücksetzen des Passworts gesendet."
+  },
+  "UpdatePassword": {
+    "title": "Passwort aktualisieren",
+    "description": "Bitte geben Sie Ihr neues Passwort ein.",
+    "passwordInputLabel": "Neues Passwort",
+    "passwordInputPlaceholder": "Geben Sie Ihr neues Passwort ein",
+    "confirmPasswordInputLabel": "Bestätigen Sie das neue Passwort",
+    "confirmPasswordInputPlaceholder": "Geben Sie Ihr neues Passwort erneut ein",
+    "updatePasswordButtonLabel": "Passwort aktualisieren",
+    "passwordsMustMatchError": "Die Passwörter müssen übereinstimmen",
+    "codeInputError": "Fehlender Code",
+    "passwordLengthError": "Das Passwort muss mindestens 8 Zeichen lang sein",
+    "passwordUppercaseError": "Das Passwort muss mindestens einen Großbuchstaben enthalten",
+    "passwordSpecialCharError": "Das Passwort muss mindestens ein Sonderzeichen enthalten"
+  },
   "confirming": "Wird bestätigt...",
   "forgotPassword": "Passwort vergessen?"
 }
