--- frontend/messages/de/Reset.json
@@ -1,8 +1,0 @@
-{
-  "metaTitle": "Passwort zurücksetzen - Coach AI",
-  "metaDescription": "Setzen Sie Ihr Passwort sicher zurück",
-  "title": "Passwort zurücksetzen",
-  "description": "Bitte geben Sie Ihre E-Mail-Adresse ein, um einen Link zum Zurücksetzen des Passworts zu erhalten.",
-  "sendResetEmailButtonLabel": "Passwort zurücksetzen",
-  "successMessage": "Wenn ein Konto mit dieser E-Mail-Adresse existiert, wurde ein Link zum Zurücksetzen des Passworts gesendet."
-}
