--- frontend/messages/de/UpdatePassword.json
@@ -1,16 +1,0 @@
-{
-  "metaTitle": "Passwort aktualisieren - Coach AI",
-  "metaDescription": "Aktualisieren Sie Ihr Passwort sicher",
-  "title": "Passwort aktualisieren",
-  "description": "Bitte geben Sie Ihr neues Passwort ein.",
-  "passwordInputLabel": "Neues Passwort",
-  "passwordInputPlaceholder": "Geben Sie Ihr neues Passwort ein",
-  "confirmPasswordInputLabel": "Bestätigen Sie das neue Passwort",
-  "confirmPasswordInputPlaceholder": "Geben Sie Ihr neues Passwort erneut ein",
-  "updatePasswordButtonLabel": "Passwort aktualisieren",
-  "passwordsMustMatchError": "Die Passwörter müssen übereinstimmen",
-  "codeInputError": "Fehlender Code",
-  "passwordLengthError": "Das Passwort muss mindestens 8 Zeichen lang sein",
-  "passwordUppercaseError": "Das Passwort muss mindestens einen Großbuchstaben enthalten",
-  "passwordSpecialCharError": "Das Passwort muss mindestens ein Sonderzeichen enthalten"
-}
