--- frontend/messages/en/Reset.json
@@ -1,8 +1,0 @@
-{
-  "metaTitle": "Reset Password - Coach AI",
-  "metaDescription": "Reset your password securely",
-  "title": "Reset Your Password",
-  "description": "Please enter your email address to receive a password reset link.",
-  "sendResetEmailButtonLabel": "Reset Password",
-  "successMessage": "If an account with that email exists, a password reset link has been sent."
-}
