--- frontend/messages/en/UpdatePassword.json
@@ -1,16 +1,0 @@
-{
-  "metaTitle": "Update Password - Coach AI",
-  "metaDescription": "Update your password securely",
-  "title": "Update Your Password",
-  "description": "Please enter your new password.",
-  "passwordInputLabel": "New Password",
-  "passwordInputPlaceholder": "Enter your new password",
-  "confirmPasswordInputLabel": "Confirm New Password",
-  "confirmPasswordInputPlaceholder": "Re-enter your new password",
-  "updatePasswordButtonLabel": "Update Password",
-  "passwordsMustMatchError": "Passwords must match",
-  "codeInputError": "Missing code",
-  "passwordLengthError": "Password must be at least 8 characters long",
-  "passwordUppercaseError": "Password must contain at least one uppercase letter",
-  "passwordSpecialCharError": "Password must contain at least one special character"
-}
