+++ frontend/src/app/[locale]/(auth)/login/components/ConfirmationForm.tsx
@@ -16,6 +16,7 @@ import { useSearchParams, useRouter } from 'next/navigation';
 import { useEffect, useState, useRef } from 'react';
 import { useForm } from 'react-hook-form';
 import z from 'zod';
+import { ModalWrapper } from './ModelWrapper';
 
 export default function ConfirmationForm() {
   const t = useTranslations('Login');
@@ -113,9 +114,8 @@ export default function ConfirmationForm() {
   };
 
   return (
-    <div className="fixed inset-0 flex items-center justify-center z-50">
-      <div className="absolute inset-0 bg-black/30 backdrop-blur-sm" />
-      <Card className="relative w-[90%] max-w-sm border-0 bg-white animate-in fade-in zoom-in duration-200">
+    <ModalWrapper>
+      <Card className="border-0 bg-white animate-in fade-in zoom-in duration-200">
         <Form {...form}>
           <form onSubmit={form.handleSubmit(verifyOTP)}>
             <CardContent className="space-y-4 p-4">
@@ -193,6 +193,6 @@ export default function ConfirmationForm() {
           </form>
         </Form>
       </Card>
-    </div>
+    </ModalWrapper>
   );
 }
