+++ frontend/src/lib/supabase/middleware.ts
@@ -18,8 +18,8 @@ const stripLocaleFromPath = (path: string): string => {
   return path;
 };
 
-const publicRoutes = ['/', '/terms', '/privacy', '/reset', '/update-password'];
-const authRoutes = ['/login', '/confirm'];
+const publicRoutes = ['/', '/terms', '/privacy'];
+const authRoutes = ['/login'];
 const adminRoutes = ['/admin'];
 
 export async function authMiddleware(
